+++ docs/changelog.rst
@@ -27,6 +27,8 @@ License
   (`#459 <https://github.com/useblocks/sphinxcontrib-needs/issues/459>`_)
 * Bugfix: Changed :ref:`needs_external_needs` getting from URL was using parameter related to local file.
   (`#458 <https://github.com/useblocks/sphinxcontrib-needs/issues/458>`_)
+* Improvement: Supports :ref:`filter_func` for :ref:`needpie`
+  (`#400 <https://github.com/useblocks/sphinxcontrib-needs/issues/400>`_)
 
 0.7.4
 -----